Load the library (and its dependencies) from its path rather than potentially the applications
When loading the ICD its dependencies will use the standard search path which uses the applications path first which means the incorrect (or even malicious) modules could be loaded. This change restricts it by using the modules path. This will fall back to the original logic when paths are not absolute.