Disable MASM assembly in Vulkan Loader on Windows
Vulkan Loader fails to build when the MASM assembler is configured
correctly on Windows because the generation of gen_defines.asm by
asm_offsets doesn't run and the other assembly sources depend on this.
The MASM assembler needs to be configured in order to get Sentry's
integration of crashpad working.

Fixed by hard-coding CMAKE_ASM_MASM_COMPILER_WORKS false even when MASM
is configured correctly.  This keeps the existing behavior.